Allow space for a bootloader installed with dd
Implement bootoffset to explcitly allow space for
a bootloader to be installed in front of the
first partition. Add support for cubietruck in
the example to dd the Debian uboot image by adding
a parameter to the customise script call.
Increase size of boot partition to make upgrades easier.
Allow the partition type to be specified.
Use mib units as these are converted to 1024 instead of 1000.

+++ examples/cubietruck-customise.sh
@@ -3,10 +3,22 @@
 set -e
 
 rootdir=$1
+image=$2
 
-# u-boot needs to be dd'd to the partition
-#cp /usr/lib/u-boot/Cubietruck/uboot.elf /boot/
-#cp /usr/lib/u-boot/Cubietruck/u-boot-sunxi-with-spl.bin /boot/
+if [ -z "${image}" ]; then
+	echo "Image not specified"
+	exit
+fi
+
+if [ ! -f '/usr/lib/u-boot/Cubietruck/u-boot-sunxi-with-spl.bin' ]; then
+	echo "Unable to find cubietruck u-boot file"
+	exit
+fi
+
+# u-boot needs to be dd'd to the device, not a partition
+# but kpartx does not setup the device, just the partitions
+
+dd if=/usr/lib/u-boot/Cubietruck/u-boot-sunxi-with-spl.bin of=${image} bs=1k seek=8
 
 mkdir -p $rootdir/boot/dtbs
 cp $rootdir/usr/lib/linux-image-*-armmp/* $rootdir/boot/dtbs
